+++ features/overlays/crimsonIsleTracker.js
@@ -3,7 +3,7 @@ import * as triggers from '../../constants/triggers';
 import * as seaCreatures from '../../constants/seaCreatures';
 import { persistentData } from "../../data/data";
 import { overlayCoordsData } from "../../data/overlayCoords";
-import { BOLD, GOLD, LIGHT_PURPLE, RED, WHITE, GRAY, DARK_RED, DARK_GRAY, RESET } from "../../constants/formatting";
+import { BOLD, GOLD, LIGHT_PURPLE, RED, WHITE, GRAY, DARK_GRAY, RESET, AQUA } from "../../constants/formatting";
 import { getLastFishingHookSeenAt, getWorldName, isInSkyblock } from "../../utils/playerState";
 import { formatDate, formatNumberWithSpaces, formatTimeElapsedBetweenDates, isDoubleHook } from "../../utils/common";
 import { CRIMSON_ISLE } from "../../constants/areas";
@@ -11,12 +11,26 @@ import { MEME_SOUND_MODE, NORMAL_SOUND_MODE, SAD_TROMBONE_SOUND_SOURCE } from ".
 import { createButtonsDisplay, toggleButtonsDisplay } from "../../utils/overlays";
 import { registerIf } from "../../utils/registers";
 
+const isInHotspot = true; // TODO
+
 triggers.REGULAR_CRIMSON_CATCH_TRIGGERS.forEach(entry => {
     registerIf(
         register("Chat", (event) => trackRegularSeaCreatureCatch()).setCriteria(entry.trigger).setContains(),
         () => settings.crimsonIsleTrackerOverlay && isInSkyblock() && getWorldName() === CRIMSON_ISLE
     );
 });
+
+const fieryScuttlerTrigger = triggers.RARE_CATCH_TRIGGERS.find(entry => entry.seaCreature === seaCreatures.FIERY_SCUTTLER);
+registerIf(
+    register("Chat", (event) => trackFieryScuttlerCatch()).setCriteria(fieryScuttlerTrigger.trigger).setContains(),
+    () => settings.crimsonIsleTrackerOverlay && isInSkyblock() && getWorldName() === CRIMSON_ISLE
+);
+
+const ragnarokTrigger = triggers.RARE_CATCH_TRIGGERS.find(entry => entry.seaCreature === seaCreatures.RAGNAROK);
+registerIf(
+    register("Chat", (event) => trackRagnarokCatch()).setCriteria(ragnarokTrigger.trigger).setContains(),
+    () => settings.crimsonIsleTrackerOverlay && isInSkyblock() && getWorldName() === CRIMSON_ISLE
+);
 
 const thunderTrigger = triggers.RARE_CATCH_TRIGGERS.find(entry => entry.seaCreature === seaCreatures.THUNDER);
 registerIf(
@@ -43,10 +57,14 @@ registerIf(
 
 register("gameUnload", () => {
     if (settings.crimsonIsleTrackerOverlay && settings.resetCrimsonIsleTrackerOnGameClosed && persistentData.crimsonIsle && (
-        persistentData.crimsonIsle.thunder.lastCatchTime ||
-        persistentData.crimsonIsle.lordJawbus.lastCatchTime ||
-        persistentData.crimsonIsle.thunder.catchesSinceLast ||
-        persistentData.crimsonIsle.lordJawbus.catchesSinceLast ||
+        persistentData.crimsonIsle.fieryScuttler?.lastCatchTime ||
+        persistentData.crimsonIsle.fieryScuttler?.catchesSinceLast ||
+        persistentData.crimsonIsle.ragnarok?.lastCatchTime ||
+        persistentData.crimsonIsle.ragnarok?.catchesSinceLast ||
+        persistentData.crimsonIsle.thunder?.lastCatchTime ||
+        persistentData.crimsonIsle.thunder?.catchesSinceLast ||
+        persistentData.crimsonIsle.lordJawbus?.lastCatchTime ||
+        persistentData.crimsonIsle.lordJawbus?.catchesSinceLast ||
         persistentData.crimsonIsle.radioactiveVials.count
     )) {
         resetCrimsonIsleTracker(true);
@@ -66,11 +84,7 @@ export function resetCrimsonIsleTracker(isConfirmed) {
             return;
         }
     
-        persistentData.crimsonIsle = {
-            thunder: { catchesSinceLast: 0, lastCatchTime: null, catchesHistory: [], averageCatches: 0 },
-            lordJawbus: { catchesSinceLast: 0, lastCatchTime: null, catchesHistory: [], averageCatches: 0 },
-            radioactiveVials: { count: 0, lordJawbusCatchesSinceLast: 0, dropsHistory: [] }
-        };
+        persistentData.crimsonIsle = getDefaultObject();
         persistentData.save();
 
         ChatLib.chat(`${GOLD}[FeeshNotifier] ${WHITE}Crimson Isle tracker was reset.`);    
@@ -124,14 +138,114 @@ export function setRadioactiveVials(count, lastOn) {
     }
 }
 
+function initMissingPersistentData() { // Init data added later and missing for users with previous version
+    if (!persistentData.crimsonIsle) return;
+    if (!persistentData.crimsonIsle.fieryScuttler) persistentData.crimsonIsle.fieryScuttler = getDefaultSeaCreatureSectionObject();
+    if (!persistentData.crimsonIsle.ragnarok) persistentData.crimsonIsle.ragnarok = getDefaultSeaCreatureSectionObject();
+}
+
+function getDefaultSeaCreatureSectionObject() {
+    return { catchesSinceLast: 0, lastCatchTime: null, catchesHistory: [], averageCatches: 0 };
+}
+
+function getDefaultObject() {
+    return {
+        fieryScuttler: getDefaultSeaCreatureSectionObject(),
+        ragnarok: getDefaultSeaCreatureSectionObject(),
+        thunder: getDefaultSeaCreatureSectionObject(),
+        lordJawbus: getDefaultSeaCreatureSectionObject(),
+        radioactiveVials: { count: 0, lordJawbusCatchesSinceLast: 0, dropsHistory: [] }
+    };
+}
+
+function trackFieryScuttlerCatch() {
+    try {
+        if (!isInSkyblock() || getWorldName() !== CRIMSON_ISLE || !settings.crimsonIsleTrackerOverlay) {
+            return;
+        }
+
+        initMissingPersistentData();
+
+        const catchesSinceLast = persistentData.crimsonIsle.fieryScuttler.catchesSinceLast + 1;
+        const lastCatchTime = persistentData.crimsonIsle.fieryScuttler.lastCatchTime;
+        const elapsedTime = lastCatchTime ? ` ${GRAY}(${WHITE}${formatTimeElapsedBetweenDates(new Date(lastCatchTime))}${GRAY})` : '';
+
+        let catchesHistory = persistentData.crimsonIsle.fieryScuttler.catchesHistory || [];
+        catchesHistory.unshift(catchesSinceLast); // Most recent counts at the start of array
+        catchesHistory.length = Math.min(catchesHistory.length, 100); // Store last 100
+        persistentData.crimsonIsle.fieryScuttler.catchesHistory = catchesHistory;
+
+        const sumCatches = catchesHistory.reduce(function(a, b) { return a + b; }, 0);
+        persistentData.crimsonIsle.fieryScuttler.averageCatches = catchesHistory.length ? Math.round(sumCatches / catchesHistory.length) : 0;
+
+        persistentData.crimsonIsle.fieryScuttler.catchesSinceLast = 0;
+        persistentData.crimsonIsle.fieryScuttler.lastCatchTime = new Date();
+
+        persistentData.crimsonIsle.thunder.catchesSinceLast += 1;
+        persistentData.crimsonIsle.lordJawbus.catchesSinceLast += 1;
+
+        if (isInHotspot) {
+            persistentData.crimsonIsle.ragnarok.catchesSinceLast += 1;
+        }
+
+        persistentData.save();
+
+        ChatLib.chat(`${GOLD}[FeeshNotifier] ${GRAY}It took ${WHITE}${catchesSinceLast} ${GRAY}${catchesSinceLast === 1 ? 'catch' : 'catches'}${elapsedTime} to get the ${GOLD}Fiery Scuttler${GRAY}.`);
+    } catch (e) {
+		console.error(e);
+		console.log(`[FeeshNotifier] Failed to track Fiery Scuttler catch.`);
+	}
+}
+
+function trackRagnarokCatch() {
+    try {
+        if (!isInSkyblock() || getWorldName() !== CRIMSON_ISLE || !settings.crimsonIsleTrackerOverlay) {
+            return;
+        }
+
+        initMissingPersistentData();
+
+        const catchesSinceLast = persistentData.crimsonIsle.ragnarok.catchesSinceLast + 1;
+        const lastCatchTime = persistentData.crimsonIsle.ragnarok.lastCatchTime;
+        const elapsedTime = lastCatchTime ? ` ${GRAY}(${WHITE}${formatTimeElapsedBetweenDates(new Date(lastCatchTime))}${GRAY})` : '';
+
+        let catchesHistory = persistentData.crimsonIsle.ragnarok.catchesHistory || [];
+        catchesHistory.unshift(catchesSinceLast); // Most recent counts at the start of array
+        catchesHistory.length = Math.min(catchesHistory.length, 100); // Store last 100
+        persistentData.crimsonIsle.ragnarok.catchesHistory = catchesHistory;
+
+        const sumCatches = catchesHistory.reduce(function(a, b) { return a + b; }, 0);
+        persistentData.crimsonIsle.ragnarok.averageCatches = catchesHistory.length ? Math.round(sumCatches / catchesHistory.length) : 0;
+
+        persistentData.crimsonIsle.ragnarok.catchesSinceLast = 0;
+        persistentData.crimsonIsle.ragnarok.lastCatchTime = new Date();
+
+        persistentData.crimsonIsle.thunder.catchesSinceLast += 1;
+        persistentData.crimsonIsle.lordJawbus.catchesSinceLast += 1;
+
+        if (isInHotspot) {
+            persistentData.crimsonIsle.fieryScuttler.catchesSinceLast += 1;
+        }
+
+        persistentData.save();
+
+        ChatLib.chat(`${GOLD}[FeeshNotifier] ${GRAY}It took ${WHITE}${catchesSinceLast} ${GRAY}${catchesSinceLast === 1 ? 'catch' : 'catches'}${elapsedTime} to get the ${LIGHT_PURPLE}Ragnarok${GRAY}.`);
+    } catch (e) {
+		console.error(e);
+		console.log(`[FeeshNotifier] Failed to track Ragnarok catch.`);
+	}
+}
+
 function trackThunderCatch() {
     try {
         if (!isInSkyblock() || getWorldName() !== CRIMSON_ISLE || !settings.crimsonIsleTrackerOverlay) {
             return;
         }
 
-        const catchesSinceLast = persistentData.crimsonIsle.thunder.catchesSinceLast;
-        const lastCatchTime = persistentData.crimsonIsle.thunder.lastCatchTime;
+        initMissingPersistentData();
+
+        const catchesSinceLast = persistentData.crimsonIsle.thunder.catchesSinceLast + 1;
+        const lastCatchTime = persistentData.crimsonIsle.thunder.lastCatchTime || null;
         const elapsedTime = lastCatchTime ? ` ${GRAY}(${WHITE}${formatTimeElapsedBetweenDates(new Date(lastCatchTime))}${GRAY})` : '';
 
         let catchesHistory = persistentData.crimsonIsle.thunder.catchesHistory || [];
@@ -146,6 +260,11 @@ function trackThunderCatch() {
         persistentData.crimsonIsle.thunder.lastCatchTime = new Date();
 
         persistentData.crimsonIsle.lordJawbus.catchesSinceLast += 1;
+
+        if (isInHotspot) {
+            persistentData.crimsonIsle.fieryScuttler.catchesSinceLast += 1;
+            persistentData.crimsonIsle.ragnarok.catchesSinceLast += 1;    
+        }
 
         persistentData.save();
 
@@ -162,9 +281,9 @@ function trackLordJawbusCatch() {
             return;
         }
 
-        const isDoubleHooked = isDoubleHook();
+        initMissingPersistentData();
 
-        const catchesSinceLast = persistentData.crimsonIsle.lordJawbus.catchesSinceLast;
+        const catchesSinceLast = persistentData.crimsonIsle.lordJawbus.catchesSinceLast + 1;
         const lastCatchTime = persistentData.crimsonIsle.lordJawbus.lastCatchTime;
         const elapsedTime = lastCatchTime ? ` ${GRAY}(${WHITE}${formatTimeElapsedBetweenDates(new Date(lastCatchTime))}${GRAY})` : '';
 
@@ -181,6 +300,12 @@ function trackLordJawbusCatch() {
 
         persistentData.crimsonIsle.thunder.catchesSinceLast += 1;
 
+        if (isInHotspot) {
+            persistentData.crimsonIsle.fieryScuttler.catchesSinceLast += 1;
+            persistentData.crimsonIsle.ragnarok.catchesSinceLast += 1;    
+        }
+
+        const isDoubleHooked = isDoubleHook();
         const valueToAdd = isDoubleHooked ? 2 : 1;
         let lordJawbusCatchesSinceLast = persistentData.crimsonIsle.radioactiveVials.lordJawbusCatchesSinceLast || 0;
         lordJawbusCatchesSinceLast += valueToAdd;
@@ -201,12 +326,19 @@ function trackRegularSeaCreatureCatch() {
             return;
         }
 
+        initMissingPersistentData();
+
         persistentData.crimsonIsle.thunder.catchesSinceLast += 1;
         persistentData.crimsonIsle.lordJawbus.catchesSinceLast += 1;
+
+        if (isInHotspot) {
+            persistentData.crimsonIsle.fieryScuttler.catchesSinceLast += 1;
+            persistentData.crimsonIsle.ragnarok.catchesSinceLast += 1;    
+        }
+
         persistentData.save();
 
-        if (persistentData.crimsonIsle.lordJawbus.catchesSinceLast &&
-            persistentData.crimsonIsle.lordJawbus.catchesSinceLast % 1000 === 0) {
+        if (persistentData.crimsonIsle.lordJawbus.catchesSinceLast && persistentData.crimsonIsle.lordJawbus.catchesSinceLast % 1000 === 0) {
             Client.showTitle('', `${RED}No ${LIGHT_PURPLE}Lord Jawbus ${RED}for ${persistentData.crimsonIsle.lordJawbus.catchesSinceLast} catches`, 1, 45, 1);
             ChatLib.chat(`${GOLD}[FeeshNotifier] ${RED}${BOLD}Yikes! ${RESET}${RED}No ${LIGHT_PURPLE}Lord Jawbus ${RED}for ${persistentData.crimsonIsle.lordJawbus.catchesSinceLast} catches...`);
 
@@ -261,10 +393,14 @@ function renderCrimsonIsleTrackerOverlay() {
     if (!settings.crimsonIsleTrackerOverlay ||
         !persistentData.crimsonIsle ||
         (
-            !persistentData.crimsonIsle.thunder.lastCatchTime &&
-            !persistentData.crimsonIsle.lordJawbus.lastCatchTime &&
-            !persistentData.crimsonIsle.thunder.catchesSinceLast &&
-            !persistentData.crimsonIsle.lordJawbus.catchesSinceLast &&
+            !persistentData.crimsonIsle.fieryScuttler?.lastCatchTime &&
+            !persistentData.crimsonIsle.fieryScuttler?.catchesSinceLast &&
+            !persistentData.crimsonIsle.ragnarok?.lastCatchTime &&
+            !persistentData.crimsonIsle.ragnarok?.catchesSinceLast &&
+            !persistentData.crimsonIsle.thunder?.lastCatchTime &&
+            !persistentData.crimsonIsle.thunder?.catchesSinceLast &&
+            !persistentData.crimsonIsle.lordJawbus?.lastCatchTime &&
+            !persistentData.crimsonIsle.lordJawbus?.catchesSinceLast &&
             !persistentData.crimsonIsle.radioactiveVials.count
         ) ||
         !isInSkyblock() ||
@@ -276,24 +412,17 @@ function renderCrimsonIsleTrackerOverlay() {
         return;
     }
 
-    const lastCatchTimeThunder = persistentData.crimsonIsle.thunder.lastCatchTime
-        ? `${WHITE}${formatTimeElapsedBetweenDates(new Date(persistentData.crimsonIsle.thunder.lastCatchTime))} ${GRAY}(${WHITE}${formatDate(new Date(persistentData.crimsonIsle.thunder.lastCatchTime))}${GRAY})` 
-        : `${WHITE}N/A`;
-    const lastCatchTimeLordJawbus = persistentData.crimsonIsle.lordJawbus.lastCatchTime 
-        ? `${WHITE}${formatTimeElapsedBetweenDates(new Date(persistentData.crimsonIsle.lordJawbus.lastCatchTime))} ${GRAY}(${WHITE}${formatDate(new Date(persistentData.crimsonIsle.lordJawbus.lastCatchTime))}${GRAY})` 
-        : `${WHITE}N/A`;
-    const averageThunder = formatNumberWithSpaces(persistentData.crimsonIsle.thunder.averageCatches) || 'N/A';
-    const averageLordJawbus = formatNumberWithSpaces(persistentData.crimsonIsle.lordJawbus.averageCatches) || 'N/A';
+    let overlayText = `${AQUA}${BOLD}Crimson Isle tracker\n`;
+    overlayText += getFieryScuttlerOverlayText();
+    overlayText += getRagnarokOverlayText();
+    overlayText += getThunderOverlayText();
+    overlayText += getLordJawbusOverlayText();
+
     const lastTimeVial = persistentData.crimsonIsle.radioactiveVials.dropsHistory.length
         ? `${WHITE}${formatTimeElapsedBetweenDates(new Date(persistentData.crimsonIsle.radioactiveVials.dropsHistory[0].time))} ${GRAY}(${WHITE}${formatDate(new Date(persistentData.crimsonIsle.radioactiveVials.dropsHistory[0].time))}${GRAY})` 
         : `${WHITE}N/A`;
     const lordJawbusCatchesSinceLastVial = persistentData.crimsonIsle.radioactiveVials.lordJawbusCatchesSinceLast || 0;
 
-    let overlayText = `${DARK_RED}${BOLD}Crimson Isle tracker\n`;
-    overlayText += `${LIGHT_PURPLE}Thunder: ${WHITE}${formatNumberWithSpaces(persistentData.crimsonIsle.thunder.catchesSinceLast)} ${GRAY}${persistentData.crimsonIsle.thunder.catchesSinceLast !== 1 ? 'catches' : 'catch'} ago ${DARK_GRAY}(${GRAY}avg: ${WHITE}${averageThunder}${DARK_GRAY})\n`;
-    overlayText += `${GRAY}Last on: ${lastCatchTimeThunder}\n`;
-    overlayText += `${LIGHT_PURPLE}Lord Jawbus: ${WHITE}${formatNumberWithSpaces(persistentData.crimsonIsle.lordJawbus.catchesSinceLast)} ${GRAY}${persistentData.crimsonIsle.lordJawbus.catchesSinceLast !== 1 ? 'catches' : 'catch'} ago ${DARK_GRAY}(${GRAY}avg: ${WHITE}${averageLordJawbus}${DARK_GRAY})\n`;
-    overlayText += `${GRAY}Last on: ${lastCatchTimeLordJawbus}\n`;
     overlayText += `${LIGHT_PURPLE}Radioactive Vials: ${WHITE}${formatNumberWithSpaces(persistentData.crimsonIsle.radioactiveVials.count)}\n`;
     overlayText += `${GRAY}Last on: ${lastTimeVial}\n`;
     overlayText += `${GRAY}Last on: ${WHITE}${formatNumberWithSpaces(lordJawbusCatchesSinceLastVial)} ${GRAY}${lordJawbusCatchesSinceLastVial !== 1 ? 'Jawbuses' : 'Jawbus'} ago`;
@@ -304,4 +433,68 @@ function renderCrimsonIsleTrackerOverlay() {
     overlay.draw();
 
     toggleButtonsDisplay(buttonsDisplay, overlay, overlayCoordsData.crimsonIsleTrackerOverlay);
+
+    function getFieryScuttlerOverlayText() {
+        if (!isInHotspot) return '';
+
+        let overlayText = '';
+        const obj = persistentData.crimsonIsle.fieryScuttler;
+        overlayText += `${GOLD}Fiery Scuttler: ${getCatchesSinceLastOverlayText(obj)} ${getAverageCatchesOverlayText(obj)}\n`;
+        overlayText += `${getLastCatchTimeOverlayText(obj)}\n`;
+
+        return overlayText;
+    }
+
+    function getRagnarokOverlayText() {
+        if (!isInHotspot) return '';
+
+        let overlayText = '';
+        const obj = persistentData.crimsonIsle.ragnarok;
+        overlayText += `${LIGHT_PURPLE}Ragnarok: ${getCatchesSinceLastOverlayText(obj)} ${getAverageCatchesOverlayText(obj)}\n`;
+        overlayText += `${getLastCatchTimeOverlayText(obj)}\n`;
+
+        return overlayText;
+    }
+
+    function getThunderOverlayText() {
+        if (!isInHotspot) return '';
+
+        let overlayText = '';
+        const obj = persistentData.crimsonIsle.thunder;
+        overlayText += `${LIGHT_PURPLE}Thunder: ${getCatchesSinceLastOverlayText(obj)} ${getAverageCatchesOverlayText(obj)}\n`;
+        overlayText += `${getLastCatchTimeOverlayText(obj)}\n`;
+
+        return overlayText;
+    }
+
+    function getLordJawbusOverlayText() {
+        if (!isInHotspot) return '';
+
+        let overlayText = '';
+        const obj = persistentData.crimsonIsle.lordJawbus;
+        overlayText += `${LIGHT_PURPLE}Lord Jawbus: ${getCatchesSinceLastOverlayText(obj)} ${getAverageCatchesOverlayText(obj)}\n`;
+        overlayText += `${getLastCatchTimeOverlayText(obj)}\n`;
+
+        return overlayText;
+    }
+
+    function getCatchesSinceLastOverlayText(obj) {
+        const catchesSinceLast = `${WHITE}${formatNumberWithSpaces(obj?.catchesSinceLast || 0)}`;
+        const text = `${catchesSinceLast} ${GRAY}${obj?.catchesSinceLast !== 1 ? 'catches' : 'catch'} ago`;
+        return text;
+    }
+
+    function getAverageCatchesOverlayText(obj) {
+        const average = formatNumberWithSpaces(obj?.averageCatches) || 'N/A';
+        const text = `${DARK_GRAY}(${GRAY}avg: ${WHITE}${average}${DARK_GRAY})`;
+        return text;
+    }
+
+    function getLastCatchTimeOverlayText(obj) {
+        const lastCatchTime = obj?.lastCatchTime
+            ? `${WHITE}${formatTimeElapsedBetweenDates(new Date(obj.lastCatchTime))} ${GRAY}(${WHITE}${formatDate(new Date(obj.lastCatchTime))}${GRAY})` 
+            : `${WHITE}N/A`;
+        const text = `${GRAY}Last on: ${lastCatchTime}`;
+        return text;
+    }
 }